refactor: Move core UI system classes to respective modules
This commit refactors the location of several core UI system classes:

- Moved `AppRoute` from `core.uisystem` to `core.app`.
- Moved `ErrorState` from `core.uisystem` to `core.model`.
- Moved `IntentProcessor` and `Reducer` from `core.uisystem` to `core.app`.

Updated all relevant import statements across feature modules (`article`, `articles`, `audioplayer`) and the main `app` module to reflect these changes.
