+++ feature/article/src/main/kotlin/com/kesicollection/feature/article/di/ArticleModule.kt
@@ -1,6 +1,6 @@
 package com.kesicollection.feature.article.di
 
-import com.kesicollection.core.uisystem.IntentProcessorFactory
+import com.kesicollection.core.app.IntentProcessorFactory
 import com.kesicollection.feature.article.Intent
 import com.kesicollection.feature.article.UiArticleState
 import com.kesicollection.feature.article.intent.DefaultIntentProcessorFactory
@@ -25,7 +25,7 @@ annotation class ArticleDefaultDispatcher
  * dependencies will be available throughout the application's lifecycle.
  *
  * It currently provides:
- * - [IntentProcessorFactory]: A factory for creating [com.kesicollection.core.uisystem.IntentProcessor] instances.
+ * - [IntentProcessorFactory]: A factory for creating [com.kesicollection.core.app.IntentProcessor] instances.
  *   The implementation provided is [DefaultIntentProcessorFactory].
  */
 @Module
