+++ feature/article/src/main/kotlin/com/kesicollection/feature/article/intent/IntentProcessorFactory.kt
@@ -1,8 +1,8 @@
 package com.kesicollection.feature.article.intent
 
 import com.kesicollection.core.app.CrashlyticsWrapper
-import com.kesicollection.core.uisystem.IntentProcessor
-import com.kesicollection.core.uisystem.IntentProcessorFactory
+import com.kesicollection.core.app.IntentProcessor
+import com.kesicollection.core.app.IntentProcessorFactory
 import com.kesicollection.data.usecase.GetArticleByIdUseCase
 import com.kesicollection.feature.article.Intent
 import com.kesicollection.feature.article.UiArticleState
