+++ feature/articles/src/main/kotlin/com/kesicollection/articles/Navigation.kt
@@ -3,7 +3,7 @@ package com.kesicollection.articles
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.kesicollection.core.uisystem.AppRoute
+import com.kesicollection.core.app.AppRoute
 import kotlinx.serialization.Serializable
 
 @Serializable
