+++ feature/articles/src/test/kotlin/com/kesicollection/articles/intentprocessor/FetchArticlesIntentProcessorTest.kt
@@ -3,8 +3,8 @@ package com.kesicollection.articles.intentprocessor
 import com.kesicollection.articles.ArticlesErrors
 import com.kesicollection.articles.UiArticlesState
 import com.kesicollection.articles.model.asUiArticle
-import com.kesicollection.core.uisystem.ErrorState
-import com.kesicollection.core.uisystem.Reducer
+import com.kesicollection.core.model.ErrorState
+import com.kesicollection.core.app.Reducer
 import com.kesicollection.data.usecase.GetArticlesUseCase
 import com.kesicollection.data.usecase.IsArticleBookmarkedUseCase
 import com.kesicollection.testing.api.TestDoubleArticleRepository
