+++ feature/audioplayer/src/main/kotlin/com/kesicollection/feature/audioplayer/Navigation.kt
@@ -6,7 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import androidx.navigation.toRoute
-import com.kesicollection.core.uisystem.AppRoute
+import com.kesicollection.core.app.AppRoute
 import kotlinx.serialization.Serializable
 
 @Serializable
