+++ feature/audioplayer/src/main/kotlin/com/kesicollection/feature/audioplayer/UiState.kt
@@ -1,6 +1,6 @@
 package com.kesicollection.feature.audioplayer
 
-import com.kesicollection.core.uisystem.ErrorState
+import com.kesicollection.core.model.ErrorState
 
 val initialState = UiAudioPlayerState()
